requirements.txt
pandas
openpyxl
fastapi
alembic
uvicorn[standard]
sqlalchemy
python-multipart
thefuzz
python-Levenshtein
rdflib
pyarrow
pytest
pytest-asyncio
httpx<0.28.0
scholarly
numpy
scipy
# ── Sprint 1: Security ──────────────────────────────────────
python-jose[cryptography]
bcrypt
cryptography
# ── Sprint 6: Rate limiting ──────────────────────────────────
slowapi
# ── Sprint 40: Enterprise Export ─────────────────────────────
weasyprint
# ── Phase 10 Sprint 47: PowerPoint Export ────────────────────
python-pptx>=1.0.2
# ── Sprint 64: Production Databases ──────────────────────────
psycopg2-binary
pymysql
# ── Sprint 65: SSO Integration ───────────────────────────────
authlib
itsdangerous

# ── Prioridad 3: Telemetry & Monitoring
sentry-sdk[fastapi]